+++ core/Azure.Mcp.Core/src/Commands/CommandFactory.cs
@@ -22,13 +22,14 @@ public class CommandFactory
     private readonly CommandGroup _rootGroup;
     private readonly ModelsJsonContext _srcGenWithOptions;
 
+    private const string RootCommandGroupName = "azmcp";
     public const char Separator = '_';
 
     /// <summary>
     /// Mapping of tokenized command names to their <see cref="IBaseCommand" />
     /// </summary>
     private readonly Dictionary<string, IBaseCommand> _commandMap;
-    private readonly HashSet<string> _serviceAreaNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+    private readonly Dictionary<string, IAreaSetup> _commandNamesToArea = new(StringComparer.OrdinalIgnoreCase);
     private readonly ITelemetryService _telemetryService;
 
     // Add this new class inside CommandFactory
@@ -51,9 +52,9 @@ public class CommandFactory
         _serviceAreas = serviceAreas?.ToArray() ?? throw new ArgumentNullException(nameof(serviceAreas));
         _serviceProvider = serviceProvider;
         _logger = logger;
-        _rootGroup = new CommandGroup("azmcp", "Azure MCP Server");
+        _rootGroup = new CommandGroup(RootCommandGroupName, "Azure MCP Server");
         _rootCommand = CreateRootCommand();
-        _commandMap = CreateCommmandDictionary(_rootGroup, string.Empty);
+        _commandMap = CreateCommandDictionary(_rootGroup, string.Empty);
         _telemetryService = telemetryService;
         _srcGenWithOptions = new ModelsJsonContext(new JsonSerializerOptions
         {
@@ -83,7 +84,7 @@ public class CommandFactory
             {
                 if (string.Equals(group.Name, groupName, StringComparison.OrdinalIgnoreCase))
                 {
-                    Dictionary<string, IBaseCommand> commandsInGroup = CreateCommmandDictionary(group, string.Empty);
+                    var commandsInGroup = CreateCommandDictionary(group, string.Empty);
                     foreach (var (key, value) in commandsInGroup)
                     {
                         commandsFromGroups[key] = value;
@@ -98,8 +99,11 @@ public class CommandFactory
 
     private void RegisterCommandGroup()
     {
-        // Register area command groups
         var loggerFactory = _serviceProvider.GetRequiredService<ILoggerFactory>();
+
+        // Register area command groups
+        var existingGroups = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+
         foreach (var area in _serviceAreas)
         {
             if (string.IsNullOrEmpty(area.Name))
@@ -111,7 +115,7 @@ public class CommandFactory
                 throw error;
             }
 
-            if (!_serviceAreaNames.Add(area.Name))
+            if (!existingGroups.Add(area.Name))
             {
                 var matchingAreaTypes = _serviceAreas
                     .Where(x => x.Name == area.Name)
@@ -126,7 +130,23 @@ public class CommandFactory
                 throw error;
             }
 
-            area.RegisterCommands(_rootGroup, loggerFactory);
+            // Create a temporary root node to register all the area's subgroups to.
+            // Use the root to map commands back to their area.
+            var tempRoot = new CommandGroup(RootCommandGroupName, string.Empty);
+
+            area.RegisterCommands(tempRoot, loggerFactory);
+
+            var commandDictionary = CreateCommandDictionary(tempRoot, string.Empty);
+            foreach (var item in commandDictionary)
+            {
+                _commandNamesToArea.Add(item.Key, area);
+            }
+
+            foreach (var group in tempRoot.SubGroup)
+            {
+                _rootGroup.SubGroup.Add(group);
+            }
+
         }
     }
 
@@ -245,11 +265,11 @@ public class CommandFactory
     /// <summary>
     /// Finds the BaseCommand given its full command name (i.e. storage_account_list).
     /// </summary>
-    /// <param name="tokenizedName">Name of the command with prefixes.</param>
+    /// <param name="fullCommandName">Name of the command with prefixes.</param>
     /// <returns></returns>
-    public IBaseCommand? FindCommandByName(string tokenizedName)
+    public IBaseCommand? FindCommandByName(string fullCommandName)
     {
-        return _commandMap.GetValueOrDefault(tokenizedName);
+        return _commandMap.GetValueOrDefault(fullCommandName);
     }
 
     /// <summary>
@@ -263,12 +283,12 @@ public class CommandFactory
             return null;
         }
 
-        var split = tokenizedName.Split(Separator, 2);
-
-        return _serviceAreaNames.Contains(split[0]) ? split[0] : null;
+        return _commandNamesToArea.TryGetValue(tokenizedName, out var area)
+            ? area.Name
+            : null;
     }
 
-    private static Dictionary<string, IBaseCommand> CreateCommmandDictionary(CommandGroup node, string prefix)
+    internal static Dictionary<string, IBaseCommand> CreateCommandDictionary(CommandGroup node, string prefix)
     {
         var aggregated = new Dictionary<string, IBaseCommand>();
         var updatedPrefix = GetPrefix(prefix, node.Name);
@@ -289,7 +309,7 @@ public class CommandFactory
 
         foreach (var command in node.SubGroup)
         {
-            var subcommandsDictionary = CreateCommmandDictionary(command, updatedPrefix);
+            var subcommandsDictionary = CreateCommandDictionary(command, updatedPrefix);
             foreach (var item in subcommandsDictionary)
             {
                 aggregated.Add(item.Key, item.Value);
@@ -299,7 +319,7 @@ public class CommandFactory
         return aggregated;
     }
 
-    private static string GetPrefix(string currentPrefix, string additional) => string.IsNullOrEmpty(currentPrefix)
+    internal static string GetPrefix(string currentPrefix, string additional) => string.IsNullOrEmpty(currentPrefix)
         ? additional
         : currentPrefix + Separator + additional;